requirements.txt
asgiref==3.3.4
Django==3.0.4
django-mssql==1.8
django-mssql-backend==2.8.1
django-pyodbc==1.1.3
django-pyodbc-azure==2.1.0.0
django-pyodbc-azure-2019==2.1.0.0
et-xmlfile==1.0.1
gunicorn==20.0.4
jdcal==1.4.1
mysqlclient==2.0.2
numpy==1.19.4
openpyxl==3.0.6
pandas==1.2.0
psycopg2==2.8.6
pyodbc==4.0.30
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
xlrd==2.0.1
XlsxWriter==1.3.7
